--- src/components/display/Home.jsx
@@ -1,5 +1,0 @@
-const Home = () => (
-  <>Home</>
-);
-
-export default Home;
