+++ src/components/display/Recent.jsx
@@ -1,0 +1,5 @@
+const Recent = () => (
+  <>Recent</>
+);
+
+export default Recent;
